Refactoring sampling code layout
