+++ lhotse/dataset/sampling/bucketing.py
@@ -1,0 +1,273 @@
+import random
+import warnings
+from functools import reduce
+from itertools import chain
+from operator import add
+from typing import Callable, Dict, List, Tuple, Type
+
+import numpy as np
+from typing_extensions import Literal
+
+from lhotse import CutSet
+from lhotse.cut import Cut
+from lhotse.dataset.sampling.base import CutSampler
+from lhotse.dataset.sampling.single_cut import SingleCutSampler
+
+
+class BucketingSampler(CutSampler):
+    """
+    Sorts the cuts in a :class:`CutSet` by their duration and puts them into similar duration buckets.
+    For each bucket, it instantiates a simpler sampler instance, e.g. :class:`SingleCutSampler`.
+
+    It behaves like an iterable that yields lists of strings (cut IDs).
+    During iteration, it randomly selects one of the buckets to yield the batch from,
+    until all the underlying samplers are depleted (which means it's the end of an epoch).
+
+    Examples:
+
+    Bucketing sampler with 20 buckets, sampling single cuts::
+
+        >>> sampler = BucketingSampler(
+        ...    cuts,
+        ...    # BucketingSampler specific args
+        ...    sampler_type=SingleCutSampler, num_buckets=20,
+        ...    # Args passed into SingleCutSampler
+        ...    max_frames=20000
+        ... )
+
+    Bucketing sampler with 20 buckets, sampling pairs of source-target cuts::
+
+        >>> sampler = BucketingSampler(
+        ...    cuts, target_cuts,
+        ...    # BucketingSampler specific args
+        ...    sampler_type=CutPairsSampler, num_buckets=20,
+        ...    # Args passed into CutPairsSampler
+        ...    max_source_frames=20000, max_target_frames=15000
+        ... )
+    """
+
+    def __init__(
+            self,
+            *cuts: CutSet,
+            sampler_type: Type = SingleCutSampler,
+            num_buckets: int = 10,
+            bucket_method: Literal['equal_len', 'equal_duration'] = 'equal_len',
+            drop_last: bool = False,
+            seed: int = 0,
+            **kwargs: Dict
+    ) -> None:
+        """
+        BucketingSampler's constructor.
+
+        :param cuts: one or more ``CutSet`` objects.
+            The first one will be used to determine the buckets for all of them.
+            Then, all of them will be used to instantiate the per-bucket samplers.
+        :param sampler_type: a sampler type that will be created for each underlying bucket.
+        :param num_buckets: how many buckets to create.
+        :param bucket_method: how should we shape the buckets. Available options are:
+            "equal_len", where each bucket contains the same number of cuts,
+            and "equal_duration", where each bucket has the same cumulative duration
+            (but different number of cuts).
+        :param drop_last: When ``True``, we will drop all incomplete batches.
+            A batch is considered incomplete if it depleted a bucket before
+            hitting the constraint such as max_duration, max_cuts, etc.
+        :param seed: random seed for bucket selection
+        :param kwargs: Arguments used to create the underlying sampler for each bucket.
+        """
+        # Do not use the distributed capacities of the CutSampler in the top-level sampler.
+        super().__init__(
+            provide_len=all(not cs.is_lazy for cs in cuts),
+            world_size=1,
+            rank=0,
+            seed=seed
+        )
+        self.num_buckets = num_buckets
+        self.drop_last = drop_last
+        self.sampler_type = sampler_type
+        self.sampler_kwargs = kwargs
+        self.cut_sets = cuts
+        if self.cut_sets[0].is_lazy:
+            warnings.warn(
+                "Lazy CutSet detected in BucketingSampler: this is not well supported yet, "
+                "and you might experience a potentially long lag while the buckets are being created."
+            )
+
+        # Split data into buckets.
+        if bucket_method == "equal_len":
+            self.buckets = create_buckets_equal_len(*self.cut_sets, num_buckets=num_buckets)
+        elif bucket_method == "equal_duration":
+            self.buckets = create_buckets_equal_duration(
+                *self.cut_sets, num_buckets=num_buckets
+            )
+        else:
+            raise ValueError(
+                f"Unknown bucket_method: '{bucket_method}'. "
+                f"Use one of: 'equal_len' or 'equal_duration'."
+            )
+
+        # Create a separate sampler for each bucket.
+        self.bucket_samplers = [
+            self.sampler_type(*bucket_cut_sets, drop_last=drop_last, **self.sampler_kwargs)
+            for bucket_cut_sets in self.buckets
+        ]
+
+        # Initialize mutable stable.
+        self.bucket_rng = random.Random(self.seed + self.epoch)
+        self.depleted = [False] * num_buckets
+
+    def set_epoch(self, epoch: int) -> None:
+        """
+        Sets the epoch for this sampler. When :attr:`shuffle=True`, this ensures all replicas
+        use a different random ordering for each epoch. Otherwise, the next iteration of this
+        sampler will yield the same ordering.
+
+        :param epoch: Epoch number.
+        """
+        for s in self.bucket_samplers:
+            s.set_epoch(epoch)
+        super().set_epoch(epoch)
+
+    def filter(self, predicate: Callable[[Cut], bool]) -> None:
+        """
+        Add a constraint on invidual cuts that has to be satisfied to consider them.
+
+        Can be useful when handling large, lazy manifests where it is not feasible to
+        pre-filter them before instantiating the sampler.
+
+        When set, we will remove the ``__len__`` attribute on the sampler, as it is now
+        determined dynamically.
+
+        Example:
+            >>> cuts = CutSet(...)
+            ... sampler = SingleCutSampler(cuts, max_duration=100.0)
+            ... # Retain only the cuts that have at least 1s and at most 20s duration.
+            ... sampler.filter(lambda cut: 1.0 <= cut.duration <= 20.0)
+        """
+        for sampler in self.bucket_samplers:
+            sampler.filter(predicate)
+
+    def __iter__(self) -> 'BucketingSampler':
+        self.bucket_rng.seed(self.seed + self.epoch)
+        for b in self.bucket_samplers:
+            iter(b)
+        self.depleted = [False] * self.num_buckets
+        return self
+
+    def _next_batch(self):
+        while not self.is_depleted:
+            idx, sampler = self.bucket_rng.choice(self._nondepleted_samplers_with_idxs)
+            try:
+                return next(sampler)
+            except StopIteration:
+                self.depleted[idx] = True
+        raise StopIteration()
+
+    def __len__(self):
+        if self.num_batches is None:
+            self.num_batches = sum(
+                len(sampler)
+                for sampler in self.bucket_samplers
+            )
+        return self.num_batches
+
+    @property
+    def is_depleted(self) -> bool:
+        return all(self.depleted)
+
+    @property
+    def _nondepleted_samplers_with_idxs(self):
+        return [
+            (idx, bs) for idx, (bs, depleted) in
+            enumerate(zip(self.bucket_samplers, self.depleted))
+            if not depleted
+        ]
+
+    def get_report(self) -> str:
+        """Returns a string describing the statistics of the sampling process so far."""
+        total_diagnostics = reduce(add, (bucket.diagnostics for bucket in self.bucket_samplers))
+        return total_diagnostics.get_report()
+
+
+def create_buckets_equal_len(
+        *cuts: CutSet, num_buckets: int
+) -> List[Tuple[CutSet, ...]]:
+    """
+    Creates buckets of cuts with similar durations.
+    Each bucket has the same number of cuts, but different cumulative duration.
+
+    :param cuts: One or more CutSets; the input CutSets are assumed to have the same cut IDs
+        (i.e., the cuts correspond to each other and are meant to be sampled together as pairs,
+        triples, etc.).
+    :param num_buckets: The number of buckets.
+    :return: A list of CutSet buckets (or tuples of CutSet buckets, depending on the input).
+    """
+    first_cut_set = cuts[0].sort_by_duration()
+    buckets = [first_cut_set.split(num_buckets)] + [
+        cs.sort_like(first_cut_set).split(num_buckets) for cs in cuts[1:]
+    ]
+    # zip(*buckets) does:
+    # [(cs0_0, cs1_0, cs2_0), (cs0_1, cs1_1, cs2_1)] -> [(cs0_0, cs0_1), (cs1_0, cs1_1), (cs2_0, cs2_1)]
+    buckets = list(zip(*buckets))
+    return buckets
+
+
+def create_buckets_equal_duration(
+        *cuts: CutSet, num_buckets: int
+) -> List[Tuple[CutSet, ...]]:
+    """
+    Creates buckets of cuts with similar durations.
+    Each bucket has the same cumulative duration, but a different number of cuts.
+
+    :param cuts: One or more CutSets; the input CutSets are assumed to have the same cut IDs
+        (i.e., the cuts correspond to each other and are meant to be sampled together as pairs,
+        triples, etc.).
+    :param num_buckets: The number of buckets.
+    :return: A list of CutSet buckets (or tuples of CutSet buckets, depending on the input).
+    """
+    first_cut_set = cuts[0].sort_by_duration(ascending=True)
+    buckets_per_cutset = [
+        _create_buckets_equal_duration_single(
+            first_cut_set, num_buckets=num_buckets
+        )
+    ]
+    for cut_set in cuts[1:]:
+        buckets_per_cutset.append(
+            # .subset() will cause the output CutSet to have the same order of cuts as `bucket`
+            cut_set.subset(cut_ids=bucket.ids) for bucket in buckets_per_cutset[0]
+        )
+    # zip(*buckets) does:
+    # [(cs0_0, cs1_0, cs2_0), (cs0_1, cs1_1, cs2_1)] -> [(cs0_0, cs0_1), (cs1_0, cs1_1), (cs2_0, cs2_1)]
+    return list(zip(*buckets_per_cutset))
+
+
+def _create_buckets_equal_duration_single(
+        cuts: CutSet, num_buckets: int
+) -> List[CutSet]:
+    """
+    Helper method to partition a single CutSet into buckets that have the same
+    cumulative duration.
+
+    See also: :meth:`.create_buckets_from_duration_percentiles`.
+    """
+    total_duration = np.sum(c.duration for c in cuts)
+    bucket_duration = total_duration / num_buckets
+    iter_cuts = iter(cuts)
+    buckets = []
+    for bucket_idx in range(num_buckets):
+        bucket = []
+        current_duration = 0
+        try:
+            while current_duration < bucket_duration:
+                bucket.append(next(iter_cuts))
+                current_duration += bucket[-1].duration
+            # Every odd bucket, take the cut that exceeded the bucket's duration
+            # and put it in the front of the iterable, so that it goes to the
+            # next bucket instead. It will ensure that the last bucket is not too
+            # thin (otherwise all the previous buckets are a little too large).
+            if bucket_idx % 2:
+                last_cut = bucket.pop()
+                iter_cuts = chain([last_cut], iter_cuts)
+        except StopIteration:
+            assert bucket_idx == num_buckets - 1
+        buckets.append(CutSet.from_cuts(bucket))
+    return buckets
